Add files via upload
Conclusion: We have used a variety of EDA techniques to deeply understand the data. 
According to the classification reports, the Random Forest classifier, followed by the Voting classifier, Logistic Regression, and KNN, appears to have the best overall performance.
Achieved an accuracy of 84%. 
In the case of popularity predictions, it predicted that the most important factor in determining the popularity of merchandise is the ”store ratio” attribute. The Random Forest Classifier provided us with the highest test result among the other algorithms we used.